Moved to a module squirrel_api
