+++ setup.py
@@ -10,7 +10,7 @@ setup(name='squirrelapi',
         author_email='mobileoxford@oucs.ox.ac.uk',
         url='https://github.com/ox-it/python-squirrelapi',
         version='0.5',
-        py_modules=['squirrelapi'],
+        py_modules=['squirrel_api', 'squirrel_api.exceptions'],
         setup_requires=["setuptools"],
         install_requires=install_requires,
         tests_require=tests_requires,
