+++ squirrel_api/__init__.py
@@ -1,0 +1,1 @@
+from squirrel_api.api import VoicemailSuperUser, VoicemailUser, VoicemailMessage
